Function bignum::arithmatic::vector_sub

pub fn vector_sub(lhs: &Vec<u32>, rhs: &Vec<u32>) -> Result<Vec<u32>, Error>

Function that takes two vectors, each representing some arbitrary unsigned values (e.g. [1,1,1] = 111u32), and return a result representing the subtraction between the two numbers. Returns a Result as some subtraction may result in an overflow, how the error is handled is up to the caller

Error type is error::Error { error: UnsignedOverflow, cause: 'text' }

Examples

use bignum::arithmatic::vector_sub;

let a = vec![4,8,7];                                    // 487
let b = vec![1,2,0];                                    // 120
assert_eq!(vector_sub(&a, &b).unwrap(), vec![3,6,7]);   // 487 - 120 = 367
 
match vector_sub(&b, &a) {      
    Err(a) => { println!("{:?}", a); }                  // 120 - 487 = Err
    _      => { }
}