Struct bignum::error::Error

pub struct Error {
    // some fields omitted
}

The Struct returned in Result when an Error is encountered, likely in from_str when a non-numeric is found or if an empty string is passed. Contains the type of error, and the cause

Methods

impl Error

fn new<S>(e: ErrorType, s: S) -> Error where S: Into<String>