Struct bignum::BigNum

pub struct BigNum {
    // some fields omitted
}

BigNum is designed to wrap a simple struct containing a vector and offer the user to create these objects just as they would a standard type, but BigNum allows operations on an arbitrary size and is not limited to a limit.

Examples

use bignum::BigNum;
use bignum::inits::{One, Zero};

let a = BigNum::from_u32(5);    // 5
let b = &a + &One::one();       // 6

let c = &a + &b;                // 11
let d = &c - &b;                // 5
let e = &c * &d;                // 55
let h = &e / 5;                 // 11

Operators are purposely implemented on references to objects. although copyable, performing arithmatic via copies would be very inefficient compared to this implementation.

Methods

impl BigNum

fn from_u32(num: u32) -> BigNum

Create a new BigNum object from a u32 value, useful to setting the initial value of a BigNum object that lies inside the range of a u32 value.

Examples

let a = bignum::BigNum::from_u32(12345);

fn from_vec(t_raw: Vec<u32>) -> BigNum

Create a new BigNum object from a vector of u32 values

Examples

let a = bignum::BigNum::from_vec(vec![1,1,1,1,1]);

fn new(base: BigNum) -> BigNum

Contructs a new BigNum object from an existing or passed u32, useful for setting initial values such as zero, one or any other.

Examples

use bignum::inits::Zero;

let a = bignum::BigNum::new(Zero::zero());

fn to_u32(self) -> Result<u32, Error>