Struct bignum::u2048

pub struct u2048 { /* fields omitted */ }

Stack-allocated arbitrary-precision (up to certain limit) integer.

This is backed by a fixed-size array of given type ("digit"). While the array is not very large (normally some hundred bytes), copying it recklessly may result in the performance hit. Thus this is intentionally not Copy.

All operations available to bignums panic in the case of overflows. The caller is responsible to use large enough bignum types.

Methods

impl u2048

pub fn from_small(v: u32) -> u2048

Makes a bignum from one digit.

pub fn from_u64(v: u64) -> u2048

Makes a bignum from u64 value.

pub fn digits(&self) -> &[u32]

Returns the internal digits as a slice [a, b, c, ...] such that the numeric value is a + b * 2^W + c * 2^(2W) + ... where W is the number of bits in the digit type.

pub fn get_bit(&self, i: usize) -> u8

Returns the i-th bit where bit 0 is the least significant one. In other words, the bit with weight 2^i.

pub fn is_zero(&self) -> bool

Returns true if the bignum is zero.

pub fn bit_length(&self) -> usize

Returns the number of bits necessary to represent this value. Note that zero is considered to need 0 bits.

impl u2048

pub const MIN: u2048

MIN: u2048 = <u2048>::min_value()

The smallest value that can be represented by this integer type.

pub const MAX: u2048

MAX: u2048 = <u2048>::max_value()

The largest value that can be represented by this integer type.

pub const fn min_value() -> u2048

pub const fn max_value() -> u2048

pub fn from_str_radix(src: &str, radix: u32) -> Result<Self, ParseIntError>

pub fn count_ones(self) -> u32

pub fn count_zeros(self) -> u32

pub fn leading_zeros(self) -> u32

pub fn trailing_zeros(self) -> u32

pub fn rotate_left(self, n: u32) -> Self

pub fn rotate_right(self, n: u32) -> Self

pub fn swap_bytes(self) -> Self

pub fn reverse_bits(self) -> Self

pub fn from_be(x: Self) -> Self

pub fn from_le(x: Self) -> Self

pub fn to_be(self) -> Self

pub fn to_le(self) -> Self